Break up loadHints.ts and reactiveHints.ts.
